requirements.txt
django-crontab
gql
requests